+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 - Fix crash with the tracer when error happens in decoding a message during batch consuming
+- Generate schema classes for all schemas, even without a consumer/producer set
 
 # 1.14.4 - 2022-06-18